Testing HTTP server and client with a single request, it works successfully. - It fails when testing multiple requests.
