+++ Makefile
@@ -1,11 +1,14 @@
 GO_BIN_PATH = bin
 GO_SRC_PATH = simple_tcp
+GO_HTTP_SRC_PATH = http
 ROOT_PATH = $(shell pwd)
 
 NF = $(GO_NF)
 GO_NF = server client NF_A
+HTTP_NF = http_server http_client
 
 NF_GO_FILES = $(shell find $(GO_SRC_PATH)/$(%) -name "*.go" ! -name "*_test.go")
+HTTP_GO_FILES = $(shell find $(GO_HTTP_SRC_PATH)/$(%) -name "*.go" ! -name "*_test.go")
 
 VERSION = $(shell git describe --tags)
 BUILD_TIME = $(shell date -u +"%Y-%m-%dT%H:%M:%SZ")
@@ -20,11 +23,12 @@ LDFLAGS = -X github.com/free5gc/version.VERSION=$(VERSION) \
 
 .DEFAULT_GOAL: nfs
 
-nfs: $(NF)
+nfs: $(NF) http
 
 all: $(NF)
 
 $(GO_NF): % : $(GO_BIN_PATH)/%
+$(HTTP_NF): % : $(GO_BIN_PATH)/%
 
 $(GO_BIN_PATH)/%: %.go $(NF_GO_FILES)
 # $(@F): The file-within-directory part of the file name of the target.
@@ -33,3 +37,9 @@ $(GO_BIN_PATH)/%: %.go $(NF_GO_FILES)
 	CGO_LDFLAGS_ALLOW="-Wl,(--whole-archive|--no-whole-archive)" go build -ldflags "$(LDFLAGS)" -o $(ROOT_PATH)/$@ $(@F).go
 
 vpath %.go $(addprefix $(GO_SRC_PATH)/, $(GO_NF))
+
+http: $(HTTP_NF)
+$(GO_BIN_PATH)/%: $(HTTP_GO_FILES)
+	@echo "Start building $(@F)...."
+	cd $(GO_HTTP_SRC_PATH)/$(@F) && \
+	CGO_LDFLAGS_ALLOW="-Wl,(--whole-archive|--no-whole-archive)" go build -ldflags "$(LDFLAGS)" -o $(ROOT_PATH)/$@ $(@F).go
