+++ http/http_client/client_stats.csv
@@ -1,0 +1,2 @@
+NF tag, NF instance ID, NF service ID, NF assigned core, RX total,RX total dropped, TX total, TX total dropped, NF sent out, NF sent to NF,NF dropped, NF next, NF tx buffered, NF tx buffered, NF tx returned
+client, 2, 2, 4, 2, 0, 0, 0, 0, 2, 2, 0, 0, 2
