--- server/logger/logger.go
@@ -1,36 +1,0 @@
-package logger
-
-import (
-	"time"
-
-	formatter "github.com/antonfisher/nested-logrus-formatter"
-	"github.com/sirupsen/logrus"
-)
-
-var (
-	log       *logrus.Logger
-	ServerLog *logrus.Entry
-)
-
-func init() {
-	log = logrus.New()
-	log.SetReportCaller(false)
-
-	log.Formatter = &formatter.Formatter{
-		TimestampFormat: time.RFC3339,
-		TrimMessages:    true,
-		NoFieldsSpace:   true,
-		HideKeys:        true,
-		FieldsOrder:     []string{"component", "category"},
-	}
-
-	ServerLog = log.WithFields(logrus.Fields{"component": "Server", "category": "App"})
-}
-
-func SetLogLevel(level logrus.Level) {
-	log.SetLevel(level)
-}
-
-func SetReportCaller(enable bool) {
-	log.SetReportCaller(enable)
-}
